Move away git function from general.py module
The module contains only HTTP requests related functionality. The better
location for git function is scm.py module with other git functionality
such as our custom RepoID, etc.
