+++ tests/unit/package_managers/test_general.py
@@ -13,14 +13,12 @@ from requests.auth import AuthBase, HTTPBasicAuth
 
 from hermeto.core.config import get_config
 from hermeto.core.errors import FetchError
-from hermeto.core.package_managers import general
 from hermeto.core.package_managers.general import (
     _async_download_binary_file,
     async_download_files,
     download_binary_file,
     pkg_requests_session,
 )
-from tests.common_utils import GIT_REF
 
 
 @pytest.mark.parametrize("auth", [None, HTTPBasicAuth("user", "password")])
@@ -54,92 +52,6 @@ def test_download_binary_file_failed(mock_get: Any) -> None:
     expected = "Could not download http://example.org/example.tar.gz: Something went wrong"
     with pytest.raises(FetchError, match=expected):
         download_binary_file("http://example.org/example.tar.gz", "/example.tar.gz")
-
-
-@pytest.mark.parametrize(
-    "url, nonstandard_info",  # See body of function for what is standard info
-    [
-        (
-            # Standard case
-            f"git+https://github.com/monty/python@{GIT_REF}",
-            None,
-        ),
-        (
-            # Ref should be converted to lowercase
-            f"git+https://github.com/monty/python@{GIT_REF.upper()}",
-            {"ref": GIT_REF},  # Standard but be explicit about it
-        ),
-        (
-            # Repo ends with .git (that is okay)
-            f"git+https://github.com/monty/python.git@{GIT_REF}",
-            {"url": "https://github.com/monty/python.git"},
-        ),
-        (
-            # git://
-            f"git://github.com/monty/python@{GIT_REF}",
-            {"url": "git://github.com/monty/python"},
-        ),
-        (
-            # git+git://
-            f"git+git://github.com/monty/python@{GIT_REF}",
-            {"url": "git://github.com/monty/python"},
-        ),
-        (
-            # No namespace
-            f"git+https://github.com/python@{GIT_REF}",
-            {"url": "https://github.com/python", "namespace": ""},
-        ),
-        (
-            # Namespace with more parts
-            f"git+https://github.com/monty/python/and/the/holy/grail@{GIT_REF}",
-            {
-                "url": "https://github.com/monty/python/and/the/holy/grail",
-                "namespace": "monty/python/and/the/holy",
-                "repo": "grail",
-            },
-        ),
-        (
-            # Port should be part of host
-            f"git+https://github.com:443/monty/python@{GIT_REF}",
-            {"url": "https://github.com:443/monty/python", "host": "github.com:443"},
-        ),
-        (
-            # Authentication should not be part of host
-            f"git+https://user:password@github.com/monty/python@{GIT_REF}",
-            {
-                "url": "https://user:password@github.com/monty/python",
-                "host": "github.com",  # Standard but be explicit about it
-            },
-        ),
-        (
-            # Params, query and fragment should be stripped
-            f"git+https://github.com/monty/python@{GIT_REF};foo=bar?bar=baz#egg=spam",
-            {
-                # Standard but be explicit about it
-                "url": "https://github.com/monty/python",
-            },
-        ),
-        (
-            # RubyGems case
-            f"https://github.com/monty/python@{GIT_REF}",
-            {
-                # Standard but be explicit about it
-                "url": "https://github.com/monty/python",
-            },
-        ),
-    ],
-)
-def test_extract_git_info(url: str, nonstandard_info: Any) -> None:
-    """Test extraction of git info from VCS URL."""
-    info = {
-        "url": "https://github.com/monty/python",
-        "ref": GIT_REF,
-        "namespace": "monty",
-        "repo": "python",
-        "host": "github.com",
-    }
-    info.update(nonstandard_info or {})
-    assert general.extract_git_info(url) == info
 
 
 @pytest.mark.asyncio
